// CourseAdminSystem/Frontend/CourseAdminSystemAngular/src/app/model/visit.ts
import { User } from './user';
import { Toilet } from './toilet';

export interface Visit {
  VisitID?: number; //adding the ? to make it optional, as it will be auto-generated by the backend
  Userid?: number;
  Toiletid?: number;
  Time?: Date;
  Rating?: number;
  Review?: string;
  TheUser?: User;
  TheToilet?: Toilet;
}